+++ src/modules/staffMembers/store/mutations.js
@@ -1,0 +1,13 @@
+export default {
+  /**
+   * Seed staff member editor by passed model.
+   *
+   * @since [*next-version*]
+   *
+   * @param state
+   * @param model
+   */
+  setStaffMemberEditorState (state, model) {
+    state.one = model
+  },
+}